// precotex_angular/src/app/models/Tejeduria/solicitudAguja.ts
export class SolicitudAguja{
    Num_Registro: string = '';
    Fecha_Registro: string ='';
    Cod_Maq_Tejeduria: string = '';
    Cod_Ordtra: string ='';
    Tip_Trabajador: string = '';
    Cod_Tejedor: string ='';
    Nom_Tejedor: string = '';
    Cod_Tipo_Aguja: string = '';
    Tipo_Aguja: string = '';
    Talon_C1: string ='';
    Talon_C2: string = '';
    Talon_C3: string ='';
    Talon_C4: string = '';
    Talon_Pl1:  string ='';
    Talon_Pl2:  string = '';
    Cndt: string ='';    
}