// precotex_angular/src/app/models/Tejeduria/produccionRectilineo.ts
export class produccion_rectilineo{
    Fec_Produccion_Ini:string = '';
    Fec_Produccion_Fin: string = '';
    Fec_Lectura:string='';
    Cod_Ordtra: string='';
    Num_Secuencia: string='';
    Sec_Maquina: string='';
    Cod_Maquina: string='';
    Cod_Tela: string='';
    Des_Tela : string = '';
    Comb: string = '';
    Cod_Talla:string='';
    Kg_Producido: string='';
    Und_Requerido: string = '';    
    Und_Producido: string='';
    Und_Pendiente: string= '';
    Und_Ingresar: string = '';
    Und_Fallado: string='';
    Tejedor: string = '';
    Cod_Usuario:string = '';
    OC: string = '';
    Precio_Unitario: string= '';
    Total:string = '';
    Cliente:string = '';
    Id:string= '';
  }